backend/requirements.txt
Django
django-cors-headers
djangorestframework
django-filter
django-debug-toolbar
djoser
